*** empty log message ***
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@939 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/engine/xml/JRImageFactory.java
@@ -75,19 +75,19 @@ public class JRImageFactory extends JRBaseFactory
 		Byte scaleImage = (Byte)JRXmlConstants.getScaleImageMap().get(atts.getValue(ATTRIBUTE_scaleImage));
 		if (scaleImage != null)
 		{
-			image.setScaleImage(scaleImage.byteValue());
+			image.setScaleImage(scaleImage);
 		}
 
 		Byte horizontalAlignment = (Byte)JRXmlConstants.getHorizontalAlignMap().get(atts.getValue(ATTRIBUTE_hAlign));
 		if (horizontalAlignment != null)
 		{
-			image.setHorizontalAlignment(horizontalAlignment.byteValue());
+			image.setHorizontalAlignment(horizontalAlignment);
 		}
 
 		Byte verticalAlignment = (Byte)JRXmlConstants.getVerticalAlignMap().get(atts.getValue(ATTRIBUTE_vAlign));
 		if (verticalAlignment != null)
 		{
-			image.setVerticalAlignment(verticalAlignment.byteValue());
+			image.setVerticalAlignment(verticalAlignment);
 		}
 
 		String isUsingCache = atts.getValue(ATTRIBUTE_isUsingCache);
